requirements.txt
asgiref==3.4.0
bandit==1.7.0
crispy-bootstrap5==0.4
dj-database-url==0.5.0
dj-stripe==2.3.0
Django==3.2.4
django-crispy-forms==1.12.0
django-etc==1.3.0
django-social-share==2.2.1
mysqlclient @ file:///C:/Users/Abu%20joseph/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.8.5
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
